requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
easyocr==1.7.2
executing==2.2.0
filelock==3.19.1
fsspec==2025.7.0
imageio==2.37.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyclipper==1.3.0.post6
Pygments==2.19.2
pynput==1.8.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tifffile==2025.6.11
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.14.1
wcwidth==0.2.13
